Compute the White sandwich estimator of standard errors for OLS

Compute the White sandwich estimator of standard errors for ordinary least squares (OLS) regression, see White (1980) and White (1980) . For more details see also Buja et al. (2019) and Buja et al. (2019) .

comp_sandwich_qr_var(mod_fit)

Arguments

mod_fit
An lm (OLS) object

Value

(tibble) White sandwich estimator of variance for OLS regression

References

Buja A, Brown L, Berk R, George E, Pitkin E, Traskin M, Zhang K, Zhao L (2019). “Models as approximations I: consequences illustrated with linear regression.” Statist. Sci., 34(4), 523--544. ISSN 0883-4237, doi: 10.1214/18-STS693 , https://doi.org/10.1214/18-STS693.

Buja A, Brown L, Kuchibhotla AK, Berk R, George E, Zhao L (2019). “Models as approximations II: A model-free theory of parametric regression.” Statist. Sci., 34(4), 545--565. ISSN 0883-4237, doi: 10.1214/18-STS694 , https://doi.org/10.1214/18-STS694.

White H (1980). “Using least squares to approximate unknown regression functions.” Internat. Econom. Rev., 21(1), 149--170. ISSN 0020-6598, doi: 10.2307/2526245 , https://doi.org/10.2307/2526245.

White H (1980). “A heteroskedasticity-consistent covariance matrix estimator and a direct test for heteroskedasticity.” Econometrica, 48(4), 817--838. ISSN 0012-9682, doi: 10.2307/1912934 , https://doi.org/10.2307/1912934.

Examples

if (FALSE) {
n <- 1e5
X <- stats::rnorm(n, 0, 1)
y <- 2 + X * 1 + stats::rnorm(n, 0, 10)
lm_fit <- stats::lm(y ~ X)
sandwich_qr_std_err <- comp_sandwich_qr_var(lm_fit)
}